商品删除操作

+++ dkd-manage/src/main/java/com/dkd/manage/service/impl/SkuServiceImpl.java
@@ -1,7 +1,10 @@
 package com.dkd.manage.service.impl;
 
 import java.util.List;
+
+import com.dkd.common.exception.ServiceException;
 import com.dkd.common.utils.DateUtils;
+import com.dkd.manage.service.IChannelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.dkd.manage.mapper.SkuMapper;
@@ -19,6 +22,8 @@ public class SkuServiceImpl implements ISkuService
 {
     @Autowired
     private SkuMapper skuMapper;
+    @Autowired
+    private IChannelService channelService;
 
     /**
      * 查询商品管理
@@ -79,7 +84,12 @@ public class SkuServiceImpl implements ISkuService
     @Override
     public int deleteSkuBySkuIds(Long[] skuIds)
     {
-
+        //1.判断商品的id集合是否有关联货道
+        int count = channelService.countChannelBySkuIds(skuIds);
+        if (count > 0){
+            throw new ServiceException("这个商品被货道关联，无法删除");
+        }
+        //2.没有关联货道才能删除
         return skuMapper.deleteSkuBySkuIds(skuIds);
     }
